XIVY-18805 improve copilot integration test structure
Move utils into corresponding packages/classes.
