+++ ch.ivyteam.smart.core.tests.integration/src/ch/ivyteam/smart/core/aspire/AspireAPI.java
@@ -1,0 +1,40 @@
+package ch.ivyteam.smart.core.aspire;
+
+import javax.ws.rs.client.WebTarget;
+
+import ch.ivyteam.ivy.jersey.client.JerseyClientBuilder;
+import tools.jackson.databind.JsonNode;
+import tools.jackson.databind.ObjectMapper;
+
+public class AspireAPI {
+
+  private final WebTarget target;
+
+  private final static ObjectMapper MAPPER = new ObjectMapper();
+
+  private AspireAPI(WebTarget target) {
+    this.target = target;
+  }
+
+  public JsonNode spansOfResource(String resource) {
+    var request = target
+        .path("spans")
+        .queryParam("resource", resource)
+        .request();
+    try (var response = request.get()) {
+      return MAPPER.readTree(response.readEntity(String.class))
+          .get("data")
+          .get("resourceSpans").get(0)
+          .get("scopeSpans").get(0)
+          .get("spans");
+    }
+  }
+
+  public static AspireAPI create(String baseUrl) {
+    return new AspireAPI(
+        JerseyClientBuilder.create("aspire-api").toClient()
+            .target(baseUrl)
+            .path("api")
+            .path("telemetry"));
+  }
+}
