+++ ch.ivyteam.smart.core.tests.integration/src/ch/ivyteam/smart/core/copilot/CopilotContainer.java
@@ -1,0 +1,15 @@
+package ch.ivyteam.smart.core.copilot;
+
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.images.builder.ImageFromDockerfile;
+
+public class CopilotContainer extends GenericContainer<CopilotContainer> {
+
+  public CopilotContainer() {
+    super(new ImageFromDockerfile().withDockerfileFromBuilder(builder -> builder
+        .from("node:24-slim")
+        .run("npm install -g @github/copilot")
+        .build()));
+    withCommand("sleep", "infinity");
+  }
+}
